LayoutTests/platform/wpe/imported/w3c/web-platform-tests/editing/other/paste-in-list-with-inline-style.tentative-expected.txt

copy this
Original text

FAIL This tests for a bug in ReplaceSelectionCommand where styles are lost during paste. assert_equals: expected "<span style=\"font-weight: 700;\">copy this</span>" but got "Original text"
